Member 관련 로직

+++ src/main/java/durikkiri/project/service/impl/PostServiceImpl.java
@@ -1,13 +1,16 @@
 package durikkiri.project.service.impl;
 
-import durikkiri.project.entity.Category;
-import durikkiri.project.entity.Comment;
+import durikkiri.project.entity.Member;
 import durikkiri.project.entity.Image;
-import durikkiri.project.entity.Post;
 import durikkiri.project.entity.dto.HomeGetDto;
 import durikkiri.project.entity.dto.comment.CommentDto;
 import durikkiri.project.entity.dto.post.*;
+import durikkiri.project.entity.post.Category;
+import durikkiri.project.entity.post.Comment;
+import durikkiri.project.entity.post.Post;
+import durikkiri.project.exception.*;
 import durikkiri.project.repository.CommentRepository;
+import durikkiri.project.repository.MemberRepository;
 import durikkiri.project.repository.ImageRepository;
 import durikkiri.project.repository.PostRepository;
 import durikkiri.project.service.PostService;
@@ -19,7 +22,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -30,8 +33,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
-import static org.springframework.http.HttpStatus.*;
-
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
@@ -41,39 +42,38 @@ public class PostServiceImpl implements PostService {
     private final ImageRepository imageRepository;
     private final DslPostRepository dslPostRepository;
     private final CommentRepository commentRepository;
+    private final MemberRepository memberRepository;
     private final Validator validator;
     @Value("${file.dir}")
     private String fileDir;
 
-
     @Override
     @Transactional
-    public HttpStatus addPost(PostAddDto postAddDto, MultipartFile image) throws IOException {
-        HttpStatus badRequest = checkFieldValid(postAddDto);
-        if (badRequest != null) return badRequest;
-        Post savePost = postRepository.save(postAddDto.toEntity());
+    public void addPost(PostAddDto postAddDto, MultipartFile image) throws IOException {
+        String memberLoginId = SecurityContextHolder.getContext().getAuthentication().getName();
+        Member member = memberRepository.findByLoginId(memberLoginId)
+                .orElseThrow(() -> new ForbiddenException("User not found"));
+        checkFieldValid(postAddDto);
+        Post savePost = postRepository.save(postAddDto.toEntity(member));
         if (image != null) {
             Image saveImage = imageRepository.save(Image.toEntity(image, fileDir, savePost));
-            log.info("파일 저장 fullPath = {}", saveImage.getFullPath());
+            log.info("File saved fullPath = {}", saveImage.getFullPath());
             image.transferTo(new File(saveImage.getFullPath()));
         }
-        return OK;
     }
 
-    private HttpStatus checkFieldValid(PostAddDto postAddDto) {
+    private void checkFieldValid(PostAddDto postAddDto) {
         if (!postAddDto.getCategory().equals(Category.GENERAL)) {
             if (postAddDto.getFieldList().isEmpty()) {
-                return BAD_REQUEST;
+                throw new BadRequestException("Field list is empty for non-general category");
             }
             for (FieldAddDto fieldAddDto : postAddDto.getFieldList()) {
                 Set<ConstraintViolation<FieldAddDto>> violations = validator.validate(fieldAddDto);
                 if (!violations.isEmpty()) {
-                    // If any violations are found, return BAD_REQUEST
-                    return HttpStatus.BAD_REQUEST;
+                    throw new BadRequestException("Field validation failed");
                 }
             }
         }
-        return null;
     }
 
     @Override
@@ -84,32 +84,30 @@ public class PostServiceImpl implements PostService {
     @Override
     @Transactional
     public PostGetDto getPost(Long postId, boolean flag) {
-        return postRepository.findPostWithField(postId)
-                .map(post -> {
-                    if (flag) {
-                        post.updateViewCount();
-                    }
-                    post.updateStatus();
-                    return PostGetDto.toDto(post);
-                })
-                .orElse(null);
+        Post post = postRepository.findPostWithField(postId)
+                .orElseThrow(() -> new NotFoundException("Post not found"));
+        if (flag) {
+            post.updateViewCount();
+        }
+        post.updateStatus();
+        return PostGetDto.toDto(post);
     }
+    @Override
+    public List<HomeGetDto> getHome() {
+        return dslPostRepository.getHome().stream().map(HomeGetDto::toDto).toList();
 
+    }
     @Override
     @Transactional
-    public HttpStatus updatePost(Long postId, MultipartFile image, PostUpdateDto postUpdateDto) {
-        Optional<Post> findPost = postRepository.findPostWithField(postId);
-        if (findPost.isPresent()) {
-            try {
-                Post post = findPost.get();
-                post.updatePost(postUpdateDto);
-                updateImage(image, post);
-                return OK;
-            } catch (IllegalArgumentException | IOException e) {
-                return BAD_REQUEST;
-            }
+    public void updatePost(Long postId, MultipartFile image, PostUpdateDto postUpdateDto) throws IOException {
+        String memberLoginId = SecurityContextHolder.getContext().getAuthentication().getName();
+        Post post = postRepository.findPostWithField(postId)
+                .orElseThrow(() -> new NotFoundException("Post not found"));
+        if (!post.getCreatedBy().equals(memberLoginId)) {
+            throw new ForbiddenException("User not authorized to update this post");
         }
-        return NOT_FOUND;
+        post.updatePost(postUpdateDto);
+        updateImage(image, post);
     }
 
     private void updateImage(MultipartFile image, Post post) throws IOException {
@@ -120,10 +118,9 @@ public class PostServiceImpl implements PostService {
                 Image newImage = Image.toEntity(image, fileDir, post);
                 existingImage.updateImage(newImage);
                 image.transferTo(new File(newImage.getFullPath()));
-            }
-            else {
+            } else {
                 Image saveImage = imageRepository.save(Image.toEntity(image, fileDir, post));
-                log.info("파일 저장 fullPath = {}", saveImage.getFullPath());
+                log.info("File saved fullPath = {}", saveImage.getFullPath());
                 image.transferTo(new File(saveImage.getFullPath()));
             }
         }
@@ -131,65 +128,50 @@ public class PostServiceImpl implements PostService {
 
     @Override
     @Transactional
-    public HttpStatus deletePost(Long postId) {
-        Optional<Post> findPost = postRepository.findById(postId);
-        if (findPost.isPresent()) {
-            Post post = findPost.get();
-            if(post.getImage() != null) {
-                deleteImage(post.getImage());
+    public void deletePost(Long postId) {
+        String memberLoginId = SecurityContextHolder.getContext().getAuthentication().getName();
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new NotFoundException("Post not found"));
+        if (!post.getCreatedBy().equals(memberLoginId)) {
+            throw new ForbiddenException("User not authorized to delete this post");
+        }
+        if(post.getImage() != null) {
+            deleteImage(post.getImage());
+        }
+        postRepository.delete(post);
+    }
+
+    private static void deleteImage(Image image) {
+        if (image != null) {
+            String existingImagePath = image.getFullPath();
+            File existingImageFile = new File(existingImagePath);
+            if (existingImageFile.exists()) {
+                existingImageFile.delete();
             }
-            postRepository.delete(post);
-            return OK;
         }
-        return NOT_FOUND;
-    }
-
-    private static void deleteImage(Image post) {
-        String existingImagePath = post.getFullPath();
-        File existingImageFile = new File(existingImagePath);
-        if (existingImageFile.exists()) {
-            existingImageFile.delete();
-        }
-    }
-
-    @Override
-    public List<HomeGetDto> getHome() {
-        return dslPostRepository.getHome().stream().map(HomeGetDto::toDto).toList();
-
     }
 
     @Override
     @Transactional
-    public HttpStatus addComment(Long postId, CommentDto commentDto) {
-        Optional<Post> findPost = postRepository.findById(postId);
-        if (findPost.isPresent()) {
-            Post post = findPost.get();
-            post.updateComment(commentDto.toEntity(post));
-            return OK;
-
-        }
-        return NOT_FOUND;
+    public void addComment(Long postId, CommentDto commentDto) {
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new NotFoundException("Post not found"));
+        post.updateComment(commentDto.toEntity(post));
     }
 
     @Override
     @Transactional
-    public HttpStatus updateComment(Long commentId, CommentDto commentDto) {
-        Optional<Comment> findComment = commentRepository.findById(commentId);
-        if (findComment.isPresent()) {
-            findComment.get().updateComment(commentDto);
-            return OK;
-        }
-        return NOT_FOUND;
+    public void updateComment(Long commentId, CommentDto commentDto) {
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new NotFoundException("Comment not found"));
+        comment.updateComment(commentDto);
     }
 
     @Override
     @Transactional
-    public HttpStatus deleteComment(Long commentId) {
-        Optional<Comment> findComment = commentRepository.findById(commentId);
-        if (findComment.isPresent()) {
-            commentRepository.delete(findComment.get());
-            return OK;
-        }
-        return NOT_FOUND;
+    public void deleteComment(Long commentId) {
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new NotFoundException("Comment not found"));
+        commentRepository.delete(comment);
     }
 }
